add .gitignore and requirements. move relevant files to src/
