+++ .gitignore
@@ -1,0 +1,29 @@
+# Virtual environment
+
+venv/
+
+# Python cache
+
+__pycache__/
+*.pyc
+
+# MacOS
+
+.DS_Store
+
+# Model artifacts (large files)
+
+models/fake-news-bert/
+models/*.joblib
+models/*.bin
+models/*.safetensors
+models/*.json
+
+# Output files
+
+predicted_results.csv
+
+# IDE
+
+.vscode/
+.idea/
