+++ requirements.txt
@@ -1,0 +1,59 @@
+accelerate==1.13.0
+aiohappyeyeballs==2.6.1
+aiohttp==3.13.3
+aiosignal==1.4.0
+annotated-doc==0.0.4
+anyio==4.12.1
+attrs==26.1.0
+certifi==2026.2.25
+charset-normalizer==3.4.6
+click==8.3.1
+datasets==4.8.4
+dill==0.4.1
+filelock==3.25.2
+frozenlist==1.8.0
+fsspec==2026.2.0
+h11==0.16.0
+hf-xet==1.4.2
+httpcore==1.0.9
+httpx==0.28.1
+huggingface_hub==1.7.2
+idna==3.11
+Jinja2==3.1.6
+joblib==1.5.3
+markdown-it-py==4.0.0
+MarkupSafe==3.0.3
+mdurl==0.1.2
+mpmath==1.3.0
+multidict==6.7.1
+multiprocess==0.70.19
+networkx==3.6.1
+numpy==2.4.3
+packaging==26.0
+pandas==3.0.1
+propcache==0.4.1
+psutil==7.2.2
+pyarrow==23.0.1
+Pygments==2.19.2
+python-dateutil==2.9.0.post0
+PyYAML==6.0.3
+regex==2026.2.28
+requests==2.32.5
+rich==14.3.3
+safetensors==0.7.0
+scikit-learn==1.8.0
+scipy==1.17.1
+setuptools==82.0.1
+shellingham==1.5.4
+six==1.17.0
+sympy==1.14.0
+threadpoolctl==3.6.0
+tokenizers==0.22.2
+torch==2.10.0
+tqdm==4.67.3
+transformers==5.3.0
+typer==0.24.1
+typing_extensions==4.15.0
+urllib3==2.6.3
+xxhash==3.6.0
+yarl==1.23.0
